+++ SketchBuilding/Pyramid.cpp
@@ -107,7 +107,7 @@ void Pyramid::comp(const std::map<std::string, std::string>& name_map, std::vect
 		
 		// convert the coordinates
 		std::vector<glm::vec2> pts;
-		for (size_t i = _points.size() - 1; i >= 0; --i) {
+		for (int i = _points.size() - 1; i >= 0; --i) {
 			pts.push_back(glm::vec2(invMat * glm::vec4(_points[i], 0, 1)));
 		}
